refactor(core.py): Split out article info logic into article_info.py
